+++ Modules/Risk/RiskManagerView.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace 币安量化机器人.Modules.Risk
+{
+    public partial class RiskManagerView : UserControl
+    {
+        public RiskManagerView()
+        {
+            InitializeComponent();
+        }
+    }
+}